--- src/main/java/com/github/nine/entities/FileName.java
@@ -1,19 +1,0 @@
-package com.github.nine.entities;
-
-import java.nio.file.Path;
-
-public class FileName {
-    private String name;
-
-    public FileName(String name){
-        this.name = name;
-    }
-
-    public FileName(Path path){
-        this(path.toString());
-    }
-
-    public String toString(){
-        return name;
-    }
-}
