--- src/main/java/com/github/nine/traverser/PathList.java
@@ -1,26 +1,0 @@
-package com.github.nine.traverser;
-
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Stream;
-
-public class PathList {
-    private List<Path> paths;
-
-    public PathList(List<Path> paths){
-        this.paths.addAll(paths);
-    }
-
-    PathList(){
-        paths = new ArrayList<>();
-    }
-
-    Stream<Path> stream(){
-        return paths.stream();
-    }
-
-    void add(Path path){
-        paths.add(path);
-    }
-}
