+++ src/main/java/com/github/ninerules/entities/FileName.java
@@ -1,0 +1,19 @@
+package com.github.ninerules.entities;
+
+import java.nio.file.Path;
+
+public class FileName {
+    private String name;
+
+    public FileName(String name){
+        this.name = name;
+    }
+
+    public FileName(Path path){
+        this(path.toString());
+    }
+
+    public String toString(){
+        return name;
+    }
+}
