+++ src/main/java/com/github/ninerules/rules/ASTVisitorPlus.java
@@ -1,0 +1,21 @@
+package com.github.ninerules.rules;
+
+import org.eclipse.jdt.core.dom.ASTVisitor;
+import org.eclipse.jdt.core.dom.CompilationUnit;
+
+import com.github.ninerules.entities.LineNumber;
+
+public abstract class ASTVisitorPlus extends ASTVisitor{
+    private CompilationUnit unit;
+
+    @Override
+    public boolean visit(CompilationUnit node) {
+        this.unit = node;
+        return super.visit(node);
+    }
+
+    public LineNumber getLineNumber(int position){
+        int line = unit.getLineNumber(position);
+        return new LineNumber(line);
+    }
+}
