+++ src/main/java/com/github/ninerules/rules/ViolationChecker.java
@@ -1,0 +1,18 @@
+package com.github.ninerules.rules;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import com.github.ninerules.entities.FileName;
+
+public abstract class ViolationChecker extends ASTVisitorPlus{
+    private List<Violation> violations = new ArrayList<>();
+
+    public void addViolation(Violation violation){
+        violations.add(violation);
+    }
+
+    public final Results createResults(FileName fileName){
+        return new Results(fileName, violations);
+    }
+}
