+++ src/main/java/com/github/ninerules/rules/package-info.java
@@ -1,0 +1,26 @@
+/**
+ * オブジェクト指向エクセサイズの9つのルールを判定するためのパッケージ．
+ * <ol>
+ *   <li>1つのメソッドにつき，インデントは一段階とすること．</li>
+ *   <li>else句を使用しないこと．</li>
+ *   <li>すべてのプリイティブ型と文字列型をラップすること．</li>
+ *   <li>１行につき，ドットは１つまでにすること．</li>
+ *   <li>名前を省略しないこと．</li>
+ *   <li>すべてのエンティティを小さくすること．
+ *     <ul>
+ *       <li>1クラスは50行まで．</li>
+ *       <li>1つのパッケージには10ファイルまで．</li>
+ *     </ul>
+ *   </li>
+ *   <li>１つのクラスにつき，インスタンス変数は２つまでにすること．</li>
+ *   <li>ファーストクラスコレクションを使用すること．</li>
+ *   <li>Getter，Setter，プロパティを使用しないこと．</li>
+ * </ol>
+ * 
+ * The package contains rule checkers for ninerules rules of object oriented programming exercise.
+ * 
+ * @see <a href="https://www.oreilly.co.jp/books/9784873113890/">ThoughtWorksアンソロジー ---アジャイルとオブジェクト指向によるソフトウェアイノベーション---</a>，
+ * ThoughtWorks Inc.　著、株式会社オージス総研 オブジェクトの広場編集部　訳，2008年12月 発行，
+ * ISBN978-4-87311-389-0
+ */
+package com.github.ninerules.rules;
