+++ src/main/java/com/github/ninerules/traverser/FileFilter.java
@@ -1,0 +1,10 @@
+package com.github.ninerules.traverser;
+
+import java.nio.file.Path;
+import java.nio.file.attribute.BasicFileAttributes;
+
+public interface FileFilter {
+    FileFilter NULL_FILTER = (path, attributes) -> true;
+
+    boolean accept(Path path, BasicFileAttributes attributes);
+}
