+++ src/main/java/com/github/ninerules/traverser/PathList.java
@@ -1,0 +1,26 @@
+package com.github.ninerules.traverser;
+
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Stream;
+
+public class PathList {
+    private List<Path> paths;
+
+    public PathList(List<Path> paths){
+        this.paths.addAll(paths);
+    }
+
+    PathList(){
+        paths = new ArrayList<>();
+    }
+
+    Stream<Path> stream(){
+        return paths.stream();
+    }
+
+    void add(Path path){
+        paths.add(path);
+    }
+}
