--- src/test/java/com/github/nine/ThrowableConsumer.java
@@ -1,6 +1,0 @@
-package com.github.nine;
-
-@FunctionalInterface
-public interface ThrowableConsumer{
-    void consume() throws Exception;
-}
