--- src/test/java/com/github/nine/traverser/TraverserTest.java
@@ -1,45 +1,0 @@
-package com.github.nine.traverser;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.junit.Test;
-
-public class TraverserTest {
-    private Path base = Paths.get("target/test-classes/hello/src");
-
-    @Test
-    public void testBasic(){
-        Traverser traverser = new Traverser();
-        List<Path> list = traverser.stream(base)
-                .sorted()
-                .map(path -> base.relativize(path))
-                .collect(Collectors.toList());
-
-        assertThat(list.size(), is(5));
-        assertThat(list.get(0), is(Paths.get("main/java/sample/hello/HelloWorld.java")));
-        assertThat(list.get(1), is(Paths.get("main/java/sample/hello/Launcher.java")));
-        assertThat(list.get(2), is(Paths.get("main/webapp/WEB-INF/web.xml")));
-        assertThat(list.get(3), is(Paths.get("main/webapp/index.html")));
-        assertThat(list.get(4), is(Paths.get("test/java/sample/hello/AppTest.java")));
-    }
-
-    @Test
-    public void testFilter(){
-        Traverser traverser = new Traverser((name, attributes) -> name.toString().endsWith(".java"));
-        List<Path> list = traverser.stream(base)
-                .sorted()
-                .map(path -> base.relativize(path))
-                .collect(Collectors.toList());
-
-        assertThat(list.size(), is(3));
-        assertThat(list.get(0), is(Paths.get("main/java/sample/hello/HelloWorld.java")));
-        assertThat(list.get(1), is(Paths.get("main/java/sample/hello/Launcher.java")));
-        assertThat(list.get(2), is(Paths.get("test/java/sample/hello/AppTest.java")));
-    }
-}
