+++ src/test/java/com/github/ninerules/ThrowableConsumer.java
@@ -1,0 +1,6 @@
+package com.github.ninerules;
+
+@FunctionalInterface
+public interface ThrowableConsumer{
+    void consume() throws Exception;
+}
